refactor: 의존 관계에 따라 dto service로 이동
