+++ src/main/java/roomescape/service/ReservationService.java
@@ -11,8 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
+import roomescape.service.dto.ReservationRequest;
+import roomescape.service.dto.ReservationResponse;
 import roomescape.domain.ReservationRepository;
 import roomescape.domain.ReservationTimeRepository;
 import roomescape.domain.ThemeRepository;
