+++ src/main/java/roomescape/service/ThemeService.java
@@ -7,8 +7,8 @@ import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.Theme;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
+import roomescape.service.dto.ThemeRequest;
+import roomescape.service.dto.ThemeResponse;
 import roomescape.domain.ReservationRepository;
 import roomescape.domain.ThemeRepository;
