+++ src/main/java/roomescape/service/dto/ReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.service.dto;
 
 import java.time.LocalDate;
 import roomescape.domain.Reservation;
